# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления кошельками",
        "title": "Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/": {
            "get": {
                "description": "Returns a greeting message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "Root endpoint",
                "responses": {
                    "200": {
                        "description": "Hello, World!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/wallets": {
            "post": {
                "description": "Deposit or withdraw an amount from a wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Perform wallet operation",
                "parameters": [
                    {
                        "description": "Wallet operation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.WalletOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated wallet details",
                        "schema": {
                            "$ref": "#/definitions/main.Wallet"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ValidationError"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/wallets/{uuid}": {
            "get": {
                "description": "Retrieve a wallet's details using its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Get wallet by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet details",
                        "schema": {
                            "$ref": "#/definitions/main.Wallet"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.OperationType": {
            "type": "string",
            "enum": [
                "DEPOSIT",
                "WITHDRAW"
            ],
            "x-enum-varnames": [
                "Deposit",
                "Withdraw"
            ]
        },
        "main.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "main.Wallet": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
        "main.WalletOperationRequest": {
            "type": "object",
            "required": [
                "amount",
                "operationType",
                "walletId"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 0
                },
                "operationType": {
                    "enum": [
                        "DEPOSIT",
                        "WITHDRAW"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/main.OperationType"
                        }
                    ]
                },
                "walletId": {
                    "type": "string"
                }
            }
        }
    }
}